Authentication 로직 정리 (resolved #32)
